+++ client/src/game/models/shapes.ts
@@ -4,7 +4,8 @@ import type { Label } from "../shapes/interfaces";
 import type { ShapeAccess, ShapeOwner } from "../shapes/owners";
 import type { SHAPE_TYPE } from "../shapes/types";
 import type { ServerAura } from "../systems/auras/models";
-import type { Permissions, TeleportOptions } from "../systems/logic/models";
+import type { Permissions } from "../systems/logic/models";
+import type { TeleportOptions } from "../systems/logic/tp/models";
 import type { ServerTracker } from "../systems/trackers/models";
 
 import type { LayerName } from "./floor";
