+++ client/src/game/systems/logic/door/models.ts
@@ -1,0 +1,6 @@
+import type { LocalId } from "../../../id";
+
+export interface DoorRequest {
+    logic: "door";
+    door: LocalId;
+}
