+++ client/src/game/systems/logic/models.ts
@@ -1,4 +1,7 @@
-import type { Global, GlobalId, LocalId } from "../../id";
+import type { Global } from "../../id";
+
+import type { DoorRequest } from "./door/models";
+import type { TpRequest } from "./tp/models";
 
 export const DEFAULT_PERMISSIONS: Permissions = { enabled: [], request: [], disabled: ["default"] };
 export type LOGIC_TYPES = "door" | "tp";
@@ -9,16 +12,6 @@ export enum Access {
     Disabled,
 }
 
-export interface DoorRequest {
-    logic: "door";
-    door: LocalId;
-}
-export interface TpRequest {
-    logic: "tp";
-    fromZone: LocalId;
-    toZone: LocalId;
-    transfers: LocalId[];
-}
 export type RequestType = DoorRequest | TpRequest;
 export type RequestTypeResponse = Global<RequestType> & { requester: string };
 
@@ -26,10 +19,4 @@ export interface Permissions {
     enabled: string[];
     request: string[];
     disabled: string[];
-}
-
-export interface TeleportOptions {
-    permissions: Permissions;
-    location?: { id: number; spawnUuid: GlobalId };
-    immediate: boolean;
 }
